fix(announcements): re-validate stored API override against allowlist
BulletinApiClient.baseUrl was accepting any persisted override that
parsed as an HttpUrl, while the save screen enforces a loopback /
RFC1918 / *.api.tigerduck.app allowlist. A stale value left by an
older build or by adb (e.g. http://example.com/v2) would slip through
on the read path, and the subscription methods could send X-Push-Token
to that disallowed host.

The debug screen had the mirror problem: resolveEffective returned any
nonblank stored value, but the client silently falls back to the build
default when the override is malformed, so the screen reported a stale
bad pref as the active effective endpoint and gave no recovery hint.

Extract OverrideValidator out of the debug-screen file into a new
AnnouncementApiOverride.kt next to the client, plus a shared
resolveAnnouncementEndpoint() that returns the URL and a flag for
whether the override was actually applied. Both BulletinApiClient and
ApiEndpointDebugScreen now read through that resolver, so the effective
endpoint and the "active override" supporting text cannot diverge from
what the client sends. When the resolver rejects a stored override, the
screen now shows a red "rejected by the allowlist; tap Clear" message
instead of falsely claiming the override is active.
